pyproject.toml
[project]
name = "face"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "duckdb>=1.4.2",
    "fastapi>=0.122.0",
    "gdown>=5.2.0",
    "keras-facenet>=0.3.2",
    "mtcnn>=1.0.0",
    "numpy>=2.3.5",
    "opencv-python>=4.11.0.86",
    "pillow>=12.0.0",
    "python-multipart>=0.0.20",
    "scipy>=1.16.3",
    "tensorflow>=2.20.0",
    "tqdm>=4.67.1",
    "uvicorn>=0.38.0",
]
